+++ src/content/btc/index.tsx
@@ -1,20 +1,32 @@
-import browser from 'webextension-polyfill'
 import '@common/styles/inject.common'
 
-import { EXECUTE_BTC_CONTENT_SCRIPT } from '@common/constants'
 import allowlist from '@common/config/allowlist'
 
-import execute from './main'
+import { store } from '@src/store'
+import { isAllowed, getChainSimpleName } from '@common/utils'
+import {
+  genEnhancedLabels,
+  genMainAddressLabel,
+  genCopyAddressBtn
+} from '@src/content/btc/feat-scripts'
 
 export class BTCInitializer {
   static matches = allowlist.BTC_EXPLORER_MATCHES
   async init() {
-    execute()
-    browser.runtime.onMessage.addListener((message, _sender, sendResponse) => {
-      if (message === EXECUTE_BTC_CONTENT_SCRIPT) {
-        execute()
-        sendResponse()
-      }
-    })
+    const { enhancedLabels, showCopyIcon } = await store.get('options')
+    const supportWebList = await store.get('supportWebList')
+
+    const allowed = isAllowed(Object.values(supportWebList))
+
+    const chain: string | undefined = getChainSimpleName()
+
+    if (!allowed || !chain) return
+
+    if (enhancedLabels) {
+      genEnhancedLabels(chain)
+      genMainAddressLabel(chain)
+    }
+
+    if (showCopyIcon) genCopyAddressBtn()
   }
 }