+++ src/content/etherscan/page-scripts/tokentxns.tsx
@@ -4,26 +4,26 @@ import { store } from '@src/store'
 import {
   genEnhancedLabels,
   convertUTC2locale,
-  genExportTableDataBtn,
   genCopyIconBtn,
-  genTransactionHashPhalconLink
+  genTransactionHashPhalconLink,
+  hideZeroValueTransfers
 } from '../feat-scripts'
 
 const initTokentxnsPageScript = async (chain: string) => {
   const {
     enhancedLabels,
     utc2locale,
-    exportTableData,
     showCopyIcon,
-    quick2Parsers
+    quick2Parsers,
+    hideZeroValueTransfers: hideZeroValueEnabled
   } = await store.get('options')
   if (enhancedLabels) genEnhancedLabels(chain)
   if (utc2locale) convertUTC2locale(ETHERSCAN_PAGES.TOKENTXNS.name)
-  if (exportTableData)
-    genExportTableDataBtn(chain, ETHERSCAN_PAGES.TOKENTXNS.name)
   if (showCopyIcon) genCopyIconBtn(ETHERSCAN_PAGES.TOKENTXNS.name)
   if (quick2Parsers)
     genTransactionHashPhalconLink(ETHERSCAN_PAGES.TOKENTXNS.name)
+  if (hideZeroValueEnabled)
+    hideZeroValueTransfers(ETHERSCAN_PAGES.TOKENTXNS.name)
 }
 
 export default initTokentxnsPageScript